LayoutTests/accessibility/ios-simulator/has-touch-event-listener-expected.txt

button
link
Makes sure that we can check if an AXObject has touch event listener.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS button.boolAttributeValue('AXHasTouchEventListener') is true
PASS link.boolAttributeValue('AXHasTouchEventListener') is true
PASS successfullyParsed is true

TEST COMPLETE
